# Different encoding for length and oddity