As pointed out by @benaadams , length and oddity could benefit from not having them encoded as bytes. Byte addressing could result in CPU stalls so maybe by coalescing them into a single field, we could benefit from it.

After applying the changes, `Equals` got smaller, lost one branch and seems to be a bit faster.

### Benchmarks

#### Before

| Method                  | slice | length | fullKeccak | oddEnd | Mean     | Error     | StdDev    | Code Size |
|------------------------ |------ |------- |----------- |------- |---------:|----------:|----------:|----------:|
| Equals_not_equal_sliced | ?     | ?      | ?          | ?      | 2.550 ns | 0.0306 ns | 0.0286 ns |     434 B |
| Equals                  | 0     | 1      | ?          | ?      | 3.933 ns | 0.0520 ns | 0.0434 ns |     438 B |
| Equals                  | 0     | 16     | ?          | ?      | 3.932 ns | 0.0320 ns | 0.0299 ns |     438 B |
| Equals                  | 0     | 32     | ?          | ?      | 3.717 ns | 0.0421 ns | 0.0373 ns |     437 B |
| Equals                  | 0     | 48     | ?          | ?      | 3.944 ns | 0.0573 ns | 0.0536 ns |     438 B |
| Equals                  | 0     | 64     | ?          | ?      | 3.644 ns | 0.0553 ns | 0.0490 ns |     437 B |
| Equals                  | 1     | 1      | ?          | ?      | 3.648 ns | 0.0419 ns | 0.0392 ns |     437 B |
| Equals                  | 1     | 16     | ?          | ?      | 4.014 ns | 0.0757 ns | 0.0708 ns |     438 B |
| Equals                  | 1     | 32     | ?          | ?      | 3.677 ns | 0.0723 ns | 0.0676 ns |     437 B |
| Equals                  | 1     | 48     | ?          | ?      | 3.938 ns | 0.0396 ns | 0.0371 ns |     438 B |
| Hash                    | 0     | ?      | False      | ?      | 2.000 ns | 0.0122 ns | 0.0109 ns |   1,213 B |
| Hash                    | 1     | ?      | False      | ?      | 2.148 ns | 0.0153 ns | 0.0143 ns |   1,195 B |
| Hash                    | 2     | ?      | False      | ?      | 1.784 ns | 0.0300 ns | 0.0281 ns |   1,208 B |
| Append_byte_aligned     | ?     | ?      | ?          | False  | 5.614 ns | 0.0338 ns | 0.0300 ns |   1,352 B |
| Hash                    | 0     | ?      | True       | ?      | 3.603 ns | 0.0643 ns | 0.0601 ns |   1,185 B |
| Hash                    | 1     | ?      | True       | ?      | 3.012 ns | 0.0540 ns | 0.0505 ns |   1,176 B |
| Hash                    | 2     | ?      | True       | ?      | 3.603 ns | 0.0356 ns | 0.0278 ns |   1,185 B |
| Append_byte_aligned     | ?     | ?      | ?          | True   | 5.639 ns | 0.0364 ns | 0.0341 ns |   1,360 B |

#### After

| Method                  | slice | length | fullKeccak | oddEnd | Mean     | Error     | StdDev    | Code Size |
|------------------------ |------ |------- |----------- |------- |---------:|----------:|----------:|----------:|
| Equals_not_equal_sliced | ?     | ?      | ?          | ?      | 2.149 ns | 0.0457 ns | 0.0427 ns |     402 B |
| Equals                  | 0     | 1      | ?          | ?      | 3.875 ns | 0.0801 ns | 0.0750 ns |     406 B |
| Equals                  | 0     | 16     | ?          | ?      | 3.875 ns | 0.0780 ns | 0.0729 ns |     406 B |
| Equals                  | 0     | 32     | ?          | ?      | 3.952 ns | 0.0632 ns | 0.0591 ns |     407 B |
| Equals                  | 0     | 48     | ?          | ?      | 3.884 ns | 0.0819 ns | 0.0766 ns |     406 B |
| Equals                  | 0     | 64     | ?          | ?      | 3.860 ns | 0.0563 ns | 0.0527 ns |     407 B |
| Equals                  | 1     | 1      | ?          | ?      | 3.939 ns | 0.0450 ns | 0.0421 ns |     407 B |
| Equals                  | 1     | 16     | ?          | ?      | 3.910 ns | 0.0440 ns | 0.0411 ns |     407 B |
| Equals                  | 1     | 32     | ?          | ?      | 3.907 ns | 0.0460 ns | 0.0407 ns |     407 B |
| Equals                  | 1     | 48     | ?          | ?      | 3.775 ns | 0.0623 ns | 0.0553 ns |     406 B |
| Hash                    | 0     | ?      | False      | ?      | 2.069 ns | 0.0248 ns | 0.0220 ns |   1,233 B |
| Hash                    | 1     | ?      | False      | ?      | 2.191 ns | 0.0282 ns | 0.0313 ns |   1,215 B |
| Hash                    | 2     | ?      | False      | ?      | 1.777 ns | 0.0129 ns | 0.0107 ns |   1,228 B |
| Append_byte_aligned     | ?     | ?      | ?          | False  | 5.628 ns | 0.0462 ns | 0.0432 ns |   1,397 B |
| Hash                    | 0     | ?      | True       | ?      | 3.502 ns | 0.0543 ns | 0.0508 ns |   1,206 B |
| Hash                    | 1     | ?      | True       | ?      | 2.955 ns | 0.0245 ns | 0.0229 ns |   1,197 B |
| Hash                    | 2     | ?      | True       | ?      | 3.485 ns | 0.0774 ns | 0.0724 ns |   1,206 B |
| Append_byte_aligned     | ?     | ?      | ?          | True   | 5.722 ns | 0.0404 ns | 0.0378 ns |   1,397 B |